backend/requirements.txt
# ─── VOXA Backend Dependencies ───
# FastAPI + Server
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
websockets==15.0.1

# Speech-to-Text (Whisper)
faster-whisper==1.1.1
pydub==0.25.1
torch==2.5.1
transformers==4.46.3
peft==0.13.2
librosa==0.10.2

# LLM (Groq API - free tier for Mistral + LLaMA)
groq==0.25.0

# Data Layer
duckdb==1.3.0
openpyxl==3.1.5
pandas==2.2.3

# RAG (optional - ChromaDB + Sentence Transformers)
# chromadb==0.6.3
# sentence-transformers==4.1.0

# TTS
edge-tts==6.1.19

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.1.0
aiofiles==24.1.0
